refactor(edgeless): utils getAttachedPoint & getAttachedPointByDeirection & test
